refactor(analytics): restore components/analytics directory layout
Move the server component back to components/analytics/index.tsx and
colocate the client tracker as shopify-client.tsx alongside it, per the
repo's component-directory convention.
